Check if toc/cue datafile part is valid
Check if the file referred to in the datafile part of the toc/cue file
can be opened. The previous behaviour triggered a segmentation fault
when the file handle for the iso image was replaced by a null pointer if
the datafile path in the toc/cue file pointed to a non-existing path.